🔀 Develop 브랜치 merge

+++ app/models/vector.py
@@ -3,7 +3,7 @@ from dataclasses import dataclass, asdict
 
 @dataclass
 class VectorPayload:
-  standard_id: float
+  standard_id: int
   category: str
   incorrect_text: str
   proof_text: str
